requirements.txt
fastapi
pyzk
python-dotenv
psycopg2-binary
uvicorn
PyJWT